feat: PartnershipResponse nullable 수정
