+++ app/src/main/java/com/eatssu/android/data/dto/response/PartnershipRestaurantResponse.kt
@@ -1,32 +1,50 @@
 package com.eatssu.android.data.dto.response
 
+import com.eatssu.android.domain.model.PartnershipRestaurant
 import com.google.gson.annotations.SerializedName
 
 data class PartnershipRestaurantResponse(
     @SerializedName("id")
-    val id: Int,
+    val id: Int?,
     @SerializedName("partnershipType")
-    val partnershipType: String,
+    val partnershipType: String?,
     @SerializedName("storeName")
-    val storeName: String,
+    val storeName: String?,
     @SerializedName("description")
-    val description: String,
+    val description: String?,
     @SerializedName("startDate")
-    val startDate: String,
+    val startDate: String?,
     @SerializedName("endDate")
-    val endDate: String,
+    val endDate: String?,
     @SerializedName("restaurantType")
-    val restaurantType: String,
+    val restaurantType: String?,
     @SerializedName("longitude")
-    val longitude: Double,
+    val longitude: Double?,
     @SerializedName("latitude")
-    val latitude: Double,
+    val latitude: Double?,
     @SerializedName("collegeName")
-    val collegeName: String,
+    val collegeName: String?,
     @SerializedName("departmentName")
     val departmentName: String?,
     @SerializedName("partnershipLikeCount")
-    val partnershipLikeCount: Int,
+    val partnershipLikeCount: Int?,
     @SerializedName("likedByUser")
-    val likedByUser: Boolean,
+    val likedByUser: Boolean?,
 )
+
+fun PartnershipRestaurantResponse.toDomain(): PartnershipRestaurant =
+    PartnershipRestaurant(
+        id = id ?: -1,
+        partnershipType = partnershipType ?: "",
+        storeName = storeName ?: "",
+        description = description ?: "",
+        startDate = startDate ?: "",
+        endDate = endDate ?: "",
+        restaurantType = restaurantType ?: "",
+        longitude = longitude ?: 126.95661313346206,
+        latitude = latitude ?: 37.49517278813046,
+        collegeName = collegeName ?: "",
+        departmentName = departmentName ?: "",
+        partnershipLikeCount = partnershipLikeCount ?: 0,
+        likedByUser = likedByUser ?: false
+    )
